Migrate frontend settings to chrome.storage.local.
This changes settings (in the front end) to be fetched from
chrome.storage.local instead of by messaging.

This is a big win:
   - considerably less code and complexity
   - substantial reduction in the amount of messaging (the old version
     was pulling all settings on each focus event, so just entering
     insert mode or changing tabs pulled all settings anew)

settings.set was only being used in one place.  It has been replaced by
a new message "setSetting" which content scripts send to the background
page to change a settings value.