Add short delay before launching from vomnibar.
This is not entirely satisfactory.  It would be great if a delay of 0
worked (a la `nextTick`).  But it seems we need a little longer because
we need to allow the UI component messaging to complete.

An alternative would be to thread the action required through the UI
component logic.  But that's pretty ugly too.

Yet another alternative would be to have the UI Component post an "ok,
we're done, do it" message once the UI component has been removed.

Fixes #1485.