Poor-man's select frames with link hints.
This is a quick-and-dirty version of #1761.  Unlike, #1761, it does not
add a new command.  Also, the footprint is considerably smaller;  that's
largely because #1761 adds a new command and includes a (much needed)
refactoring of the link-hints code.

It turns out that this is not as useful as it might immediately appear.
The problem is that, once you're in a frame, you cannot use link hints
to get out of it.