Separate ClickableElements functions from LinkHintsMode.
Really, all of the code related to finding clickable elements is
unrelated to LinkHintsMode.  Unfortunately, it's embedded in the middle
of the class.  So, we can't use it from outside of the class.

This is a temporary restructuring of the link-hints code.  The intention
is to move the lines around eventually.  For now, however, we do it like
this to keep the diff smaller and clearer.