Move link activation to HintCoordinator.
This moves the link activation logic out of LinkHintsMode and into the
hint coordinator.

At this point, there is (almost) no DOM-specific logic left in
LinksHintMode.  It only depends an a list of "elements", each of which
has a rect property.

The main exception to this is filtered hints.  In the following commits,
we're going to leave filtered hints behind - mainly to keep the diff
shorter and cleaner.