Global link hints; fix focus issue for options page.
Here's the issue (and we may have to address this ouside of this PR).

If we put the HUD in the top frame, then the top frame grabs the focus
when the HUD is displayed.

If we open link hints with the help dialog open, then the help dialog
loses the focus, and we can't `Esc` out of it.