Add swapCase command option for marks.
Using capitals letters for global marks and lower-case letters for local
marks makes sense for vim, but not so much so for Vimium.  Vimium is not
a text editor.  Global marks are really useful as a form of quick
bookmark; local marks are (in my experience) less useful.

So, it would be useful to swap these: lower-case for global marks,
upper-case for local marks.  See #1825.

This does not warrant a new options-page option.

However, it may warrant a new command option:

    map m Marks.activateCreateMode swapCase
    map ' Marks.activateGotoMode   swapCase

The proposal would be to document this only in the wiki.

What do you think, @mrmr1993, @philc?

The risk here is options creep.  Is this a slippery slope?