Do not unregister the main/top frame.
We never unregister the main/top frame.  If the tab is navigating to
another page, then there'll be a new top frame (with the same Id) along
soon.  If the tab is closing, then we'll tidy up in the
chrome.tabs.onRemoved listener, below.  This approach avoids a
dependency on the order in which register and unregister events happens.

See comment in #2053.