Focus window when UI component closes.
On some pages, when find mode exits, Chrome drops the focus into a frame
other than the main frame, and other than the original frame (e.g. into
an advert iframe).  This makes it seem that Vimium is broken; `gF` is
required to recover.

This fixes that by explicitly focusing the originating frame.

However, this is not entirely satisfactory.  The entire UI-component
hide logic has become pretty complicated a needs to be reviewed.