Use a fresh tabs list for visitPreviousTab
With chrome://flags#enable-fast-unload enabled, the tabs.onRemoved event
sometimes doesn't fire, and so TabRecency.getTabsByRecency would return
dead tabs.

This is still a Chromium issue we need to be aware of: users with the
flag enabled will leak the memory we would normally release in the
tabs.onRemoved handlers.